% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{c_index}
\alias{c_index}
\title{Concordance Index (C-index)}
\usage{
c_index(pre, object)
}
\arguments{
\item{pre}{A numeric vector of predicted survival times.}

\item{object}{A dataframe or matrix, where the first column is observed survival time and the second column is survival status.}
}
\value{
The value of the C-index.
}
\description{
Compute the C-index in survival analysis.
}
\details{
The most frequently used evaluation metric of survival models is the concordance index (C-index).
It is a measure of rank correlation between predicted risk scores. Only uncensored observations are considered.
}
\examples{
set.seed(1)
# Simulate predicted survival times
x1 <- rnorm(100)
# Simulate observed survival times
x2 <- rnorm(100)
# Simulate observed survival status
x3 <- rep(c(1, 0), 50)
# Calculate C-index
c_index(x1, cbind(x2, x3))
}
