% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_km_weights_mat.R
\name{find_km_weights_mat}
\alias{find_km_weights_mat}
\title{Calculate Kaplan-Meier Weights}
\usage{
find_km_weights_mat(multi_dat, num_outcome)
}
\arguments{
\item{multi_dat}{A data frame containing the time and status of multi-variate survival outcomes for every subject.
The name of each column corresponding to the time or status must contain the characters "time" or "status", respectively.}

\item{num_outcome}{Number of the outcomes.}
}
\value{
A matrix of Kaplan-Meier weights for each survival outcome.
}
\description{
Produce the Kaplan-Meier weights
}
\details{
The Kaplan-Meier weights are defined as:
\deqn{ w_{n1}=\frac{d_{(1)} }{n} , w_{ni}=\frac{d_{(i)} }{n-i+1} \prod^{i-1}_{j=1} (\frac{n-j}{n-j+1} )^{d_{(j)} }, i=2,...,n.}
}
\examples{
# multi-outcome
library(dplyr)
data(simdat_mtps)
find_km_weights_mat(ymat, 2)
}
\references{
Huang J, Ma S, Xie H. Regularized estimation in the accelerated failure time model with high-dimensional covariates. Biometrics. 2006;62(3):813-820.
}
