% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.learners.R
\name{list.learners}
\alias{list.learners}
\alias{list_learners}
\title{List Available Base Learners}
\usage{
list_learners()
}
\value{
The name of all base learners provided in the package.
}
\description{
This function lists all base learners provided in the package.
}
\details{
- lm1: Linear regression
- glm1: Generalized linear models
- glmnet1: Performs k-fold cross-validation to choose the best alpha and lambda for generalized linear models via penalized maximum likelihood.
- glmnet_lasso: LASSO, lambda is chosen by k-fold cross-validation for glmnet.
- glmnet_ridge: Ridge regression, lambda is chosen by k-fold cross-validation for glmnet.
- rpart1: Regression tree
- lda1: Linear discriminant analysis
- qda1: Quadratic discriminant analysis
- KNN1: K-nearest neighbor classification, k is chosen by cross-validation.
- svm1: Support vector machine
- surv: Parametric AFT model
- ela1: Elastic Net AFT model, weights are the weight of survival time.
}
\examples{
list_learners()
}
