% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.parameter.R
\name{modify_parameter}
\alias{modify_parameter}
\title{Modify Default Parameters For Base Learner}
\usage{
modify_parameter(FUN, ...)
}
\arguments{
\item{FUN}{Method.}

\item{...}{Modified arguments.}
}
\value{
It returns a new function with modified parameters.
}
\description{
Modify default parameters for methods provided in the package.
}
\examples{
glmnet_lasso <- modify_parameter(glmnet1, alpha = 1)
glmnet_ridge <- modify_parameter(glmnet1, alpha = 0)
}
